Time: 102 ms (7.11%), Space: 202.3 MB (5.39%) - LeetHub

+++ 0907-sum-of-subarray-minimums/0907-sum-of-subarray-minimums.cpp
@@ -41,7 +41,7 @@ public:
         //     cout<<pse[i]<<" ";
             
         // }
-        long long ans = 0;
+        int ans = 0;
         for(int i=0 ; i<arr.size() ; i++){
             long long left = i - pse[i];
             long long right = nse[i] - i;
